MessageFunctions

Description

Documentation

isNumeric

Arguments

:: [Char]
String to test

-> Bool
Result

Checks if a string contains only numbers and outputs a Boolean value based on the results.

createUser

Arguments

:: Int
id of user to be created

-> User
Returns created user

Creates a user by taking id as input, intialises total messages to 0 and sets the user_messages to empty array

outputTotalMessagesPar

Arguments

:: User
Input user

-> String
Output

Gets the number of messages received by the user. Used with parMap rdeepseq to generate a list of message counts which is then displayed with mapM_

randomUser

Arguments

:: Int
RandomIndex will not choose an index that is the same as the number it was given.

-> Int
Number of Users

-> IO Int
Returns id of randomly selected user

This function randomly chooses an index associated with a user different from the one provided in the exclude_index argument, allowing a user to send a message to another user.

mapFork

Arguments

:: MVar (Seq User)
A sequence of users

-> MVar Int
Wait time MVar, to stop execution

-> MVar Int
Empty mVar to store number of messages sent

-> [Int]
No of user threads to be created

-> Int
No of Messages to be sent

-> Int
Max Number of Users

-> IO ()

This function allows for convenient generation of a large amount of threads

process

Arguments

:: Int
User id of the random user

-> MVar (Seq User)
MVar containing sequence of users

-> MVar Int
Wait time MVar, to stop execution when the maximum sent message limit is reached

-> MVar Int
MVar to store the number all the sent messages

-> Int
Total Number of messages to send

-> Int
Number of users

-> IO ()
Required output by forkIO (used to implement concurrency)

For each simulated user the following function will check if the macimum number of messages has been reached. If maximum number of messages has been reached then the user will exit the thread and initialize MVar to indicate the exit. If not then the following tasks will be performed

Choose a user randomly from the users list by using Data.Random.randomIO, which will have a different user_id from the thread_user_id.

Generate a random message to be sent by using Text.StringRandom.stringRandomIO

Message is then sent to the user and both senders and recipeint user inbox is then updated

wait for a random period of time determined by Data.Random.randomRIO.

The function can call itself repeatedly with modified values of the MVars.

atomicPutStrLn

Arguments

:: String
String to be displayed on the screen

-> IO ()
Return type of IO ()

Thread safe version of putStrLn, It uses takeMVar and putMVar functions from the Control.Concurrent.MVar module to acquire and release a lock on the standard output stream

generateRandomMessage

Arguments

:: Int
User Id of the recipient

-> Int
User Id of the Sender

-> Text
Message contentin

-> String
Date when the message was sent

-> String
Time when the message was sent

-> IO Message
Return type of IO Message

Generates a message in the specified format to be displayed on screen

stdoutLock :: MVar ()

To allow two threads to write over stdout

getCurrentDateTime

Arguments

:: IO (String, String)
A tuple of String containing current date and current time.

Function to return tuple of strings containing the current date and current time.

generateString :: IO Text

Function to generate a random string of characters that matches a given regular expression pattern

findUserById

Arguments

:: Int
UserId Of the recipient

-> [User]
List of users

-> Maybe User
Returns a user or error

Function to find a user given a userid from a list of users

displayUserInbox

Arguments

:: Int
userId of the recipient

-> Seq User
Sequence of Users

-> IO ()
Returns IO to display the user's Inbox

Function to display the user's inbox

getId

Arguments

:: User
User object

-> Int
Returns userId

Function to get userId from User

displayUserOutbox

Arguments

:: Int
userId of the recipient

-> Seq User
Sequence of Users

-> IO ()
Returns IO to display the user's Outbox

Function to display the user's outbox

displayInboxMessages

Arguments

:: User
User Object

-> IO ()
Return type IO

Function to display the Inbox Messages

displayOutboxMessages

Arguments

:: User
User Object

-> IO ()
Return type IO

Function to display the Inbox Messages