Types

Description

Documentation

data User

User Data Type, to store information about the user

Constructors

User

Fields

user_id :: Int

Unique Identifier to identify the user

messages_received :: Int

Stores the Number of Messages received by the user

received_messages :: [Message]

Stores the Messages received by the user

no_of_messages_sent :: Int

Stores the Number of Messages sent by the user

sent_messages :: [Message]

Stores the messages sent by the user

data Message

Message Data type, to store the messages sent by the users

Constructors

Message

Fields

from :: Int

Stores the user_id of the sender

to :: Int

Stores the user_id of the recipient

content :: Text

Stores the message body

time :: String

Time when the message was recieved

date :: String

Date When the message was recieved